Add recursive move item function
