+++ app/routers/v1/items/utils.py
@@ -1,0 +1,24 @@
+# Copyright (C) 2022 Indoc Research
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+def combine_item_tables(item_result: tuple) -> dict:
+    item_data = item_result[0].to_dict()
+    storage_data = item_result[1].to_dict()
+    storage_data.pop('item_id')
+    extended_data = item_result[2].to_dict()
+    extended_data.pop('item_id')
+    item_data['storage'] = storage_data
+    item_data['extended'] = extended_data
+    return item_data
